Address PHPMD core library checks
